sonarlint cleaning; remove unnecessary casts (squid:S1905)

+++ src/com/base/game/inventory/CharacterInventory.java
@@ -1329,7 +1329,7 @@ public class CharacterInventory implements Serializable {
 
 		average /= count;
 		
-		return (int) average;
+		return average;
 	}
 
 	/**
